database/sinkdb: move storage from net/raft
